pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "flake8>=7.3.0",
    "pydantic-settings>=2.10.1",
    "ruff>=0.12.11",
    "sqlalchemy>=2.0.43",
    "supabase>=2.18.1",
]
